refactor: Clean up MockNodeKeyManager pattern for tests
