+++ packages/dds/tree/src/simple-tree/toMapTree.ts
@@ -359,7 +359,7 @@ function mapToMapTree(
  * @param schema - The schema associated with the value.
  * @param nodeKeyManager - See {@link NodeKeyManager}.
  */
-export function objectToMapTree(
+function objectToMapTree(
 	data: InsertableContent,
 	schema: TreeNodeSchema,
 	nodeKeyManager: NodeKeyManager,
