+++ packages/dds/tree/src/test/feature-libraries/chunked-forest/chunkEncodingEndToEnd.ts
@@ -34,7 +34,6 @@ import {
 	ModularChangeset,
 	TreeCompressionStrategy,
 	buildChunkedForest,
-	createMockNodeKeyManager,
 	defaultSchemaPolicy,
 	fieldKindConfigurations,
 	getTreeContext,
@@ -48,6 +47,7 @@ import {
 	flexTreeViewWithContent,
 	numberSequenceRootSchema,
 } from "../../utils.js";
+import { MockNodeKeyManager } from "../../MockNodeKeyManager.js";
 
 const options = {
 	jsonValidator: typeboxValidator,
@@ -104,7 +104,7 @@ describe("End to end chunked encoding", () => {
 				schema,
 				// Note: deliberately passing an editor that doesn't have the property for schema edition; test doesn't need it
 				new MockTreeCheckout(editableForest, dummyEditor as unknown as ISharedTreeEditor),
-				createMockNodeKeyManager(),
+				new MockNodeKeyManager(),
 			);
 		}
