+++ packages/dds/tree/src/test/shared-tree/schematizingTreeView.spec.ts
@@ -11,7 +11,6 @@ import {
 	FieldKinds,
 	FlexFieldSchema,
 	SchemaBuilderBase,
-	createMockNodeKeyManager,
 	intoStoredSchema,
 } from "../../feature-libraries/index.js";
 // eslint-disable-next-line import/no-internal-modules
@@ -26,12 +25,13 @@ import { SchemaFactory, TreeConfiguration } from "../../simple-tree/index.js";
 import { toFlexConfig, toFlexSchema } from "../../simple-tree/toFlexSchema.js";
 import { disposeSymbol } from "../../util/index.js";
 import { checkoutWithContent, createTestUndoRedoStacks, insert } from "../utils.js";
+import { MockNodeKeyManager } from "../MockNodeKeyManager.js";
 
 const schema = new SchemaFactory("com.example");
 const config = new TreeConfiguration(schema.number, () => 5);
 const configGeneralized = new TreeConfiguration([schema.number, schema.string], () => 6);
-const flexConfig = toFlexConfig(config, createMockNodeKeyManager());
-const flexConfigGeneralized = toFlexConfig(configGeneralized, createMockNodeKeyManager());
+const flexConfig = toFlexConfig(config, new MockNodeKeyManager());
+const flexConfigGeneralized = toFlexConfig(configGeneralized, new MockNodeKeyManager());
 
 // Schema for tree that must always be empty.
 const emptySchema = new SchemaBuilderBase(FieldKinds.required, {
@@ -49,7 +49,7 @@ describe("SchematizingSimpleTreeView", () => {
 			initialTree: undefined,
 		};
 		const checkout = checkoutWithContent(emptyContent);
-		const view = new SchematizingSimpleTreeView(checkout, config, createMockNodeKeyManager());
+		const view = new SchematizingSimpleTreeView(checkout, config, new MockNodeKeyManager());
 
 		const error: SchematizeError | undefined = view.error;
 		assert(error instanceof SchematizeError);
@@ -63,7 +63,7 @@ describe("SchematizingSimpleTreeView", () => {
 
 	it("Open and close existing document", () => {
 		const checkout = checkoutWithContent(flexConfig);
-		const view = new SchematizingSimpleTreeView(checkout, config, createMockNodeKeyManager());
+		const view = new SchematizingSimpleTreeView(checkout, config, new MockNodeKeyManager());
 		assert.equal(view.error, undefined);
 		const root = view.root;
 		assert.equal(root, 5);
@@ -92,7 +92,7 @@ describe("SchematizingSimpleTreeView", () => {
 
 	it("Modify root", () => {
 		const checkout = checkoutWithContent(flexConfig);
-		const view = new SchematizingSimpleTreeView(checkout, config, createMockNodeKeyManager());
+		const view = new SchematizingSimpleTreeView(checkout, config, new MockNodeKeyManager());
 		view.events.on("rootChanged", () => log.push(["rootChanged", view.error ?? view.root]));
 		view.events.on("afterBatch", () => log.push(["afterBatch", view.root]));
 		assert.equal(view.root, 5);
@@ -109,7 +109,7 @@ describe("SchematizingSimpleTreeView", () => {
 
 	it("Schema becomes incompatible then comparable", () => {
 		const checkout = checkoutWithContent(flexConfig);
-		const view = new SchematizingSimpleTreeView(checkout, config, createMockNodeKeyManager());
+		const view = new SchematizingSimpleTreeView(checkout, config, new MockNodeKeyManager());
 		assert.equal(view.root, 5);
 		const log: [string, unknown][] = [];
 		view.events.on("rootChanged", () => log.push(["rootChanged", view.error ?? view.root]));
@@ -147,7 +147,7 @@ describe("SchematizingSimpleTreeView", () => {
 		const view = new SchematizingSimpleTreeView(
 			checkout,
 			configGeneralized,
-			createMockNodeKeyManager(),
+			new MockNodeKeyManager(),
 		);
 		const log: [string, unknown][] = [];
 		view.events.on("rootChanged", () => log.push(["rootChanged", view.error ?? view.root]));
@@ -172,7 +172,7 @@ describe("SchematizingSimpleTreeView", () => {
 
 	it("Open incompatible document", () => {
 		const checkout = checkoutWithContent(flexConfigGeneralized);
-		const view = new SchematizingSimpleTreeView(checkout, config, createMockNodeKeyManager());
+		const view = new SchematizingSimpleTreeView(checkout, config, new MockNodeKeyManager());
 
 		const error: SchematizeError | undefined = view.error;
 		assert(error instanceof SchematizeError);
@@ -196,7 +196,7 @@ describe("SchematizingSimpleTreeView", () => {
 			initialTree: undefined,
 		};
 		const checkout = checkoutWithContent(emptyContent);
-		const view = new SchematizingSimpleTreeView(checkout, config, createMockNodeKeyManager());
+		const view = new SchematizingSimpleTreeView(checkout, config, new MockNodeKeyManager());
 
 		const { undoStack, redoStack } = createTestUndoRedoStacks(view.events);
 
@@ -214,7 +214,7 @@ describe("SchematizingSimpleTreeView", () => {
 		const sf = new SchemaFactory(undefined);
 		class TestObject extends sf.object("TestObject", { value: sf.number }) {}
 		const treeContent = new TreeConfiguration(TestObject, () => new TestObject({ value: 3 }));
-		const nodeKeyManager = createMockNodeKeyManager();
+		const nodeKeyManager = new MockNodeKeyManager();
 		const view = new SchematizingSimpleTreeView(
 			checkoutWithContent(toFlexConfig(treeContent, nodeKeyManager)),
 			treeContent,
